+++ src/main/java/pl/arcsoftware/carcoderepo/models/CompanyEntity.java
@@ -12,7 +12,7 @@ public class CompanyEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
     @Basic
     @Column(name = "name", nullable = true, length = 45)
     private String name;
@@ -28,11 +28,11 @@ public class CompanyEntity {
     @OneToMany(mappedBy = "companyByCompanyId")
     private Collection<MaintenanceEntity> maintenancesById;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public CompanyEntity setId(Integer id) {
+    public CompanyEntity setId(Long id) {
         this.id = id;
         return this;
     }
