+++ src/main/java/pl/arcsoftware/carcoderepo/security/services/UserDetailsImpl.java
@@ -14,14 +14,14 @@ import java.util.Objects;
 public class UserDetailsImpl implements UserDetails {
 
     private static final long serialVersionUID = 1L;
-    private Integer id;
+    private Long id;
     private String username;
     private String email;
     @JsonIgnore
     private String password;
     private Collection<? extends GrantedAuthority> authorities;
 
-    public UserDetailsImpl(Integer id, String username, String email, String password, Collection<? extends GrantedAuthority> authorities) {
+    public UserDetailsImpl(Long id, String username, String email, String password, Collection<? extends GrantedAuthority> authorities) {
         this.id = id;
         this.username = username;
         this.email = email;
@@ -39,11 +39,11 @@ public class UserDetailsImpl implements UserDetails {
                 authorities);
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public UserDetailsImpl setId(Integer id) {
+    public UserDetailsImpl setId(Long id) {
         this.id = id;
         return this;
     }
